+++ BE/src/main/java/com/codesquad/issueTracker/issue/domain/repository/IssueRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.codesquad.issueTracker.issue.domain.repository;
 import java.util.List;
 
 import com.codesquad.issueTracker.issue.application.dto.IssueCoverResponse;
-import com.codesquad.issueTracker.issue.presentation.dto.FilterCondition;
+import com.codesquad.issueTracker.issue.application.dto.FilterCondition;
 
 public interface IssueRepositoryCustom {
     List<IssueCoverResponse> search(FilterCondition condition, Long userId);
