+++ BE/src/main/java/com/codesquad/issueTracker/issue/domain/repository/IssueRepositoryCustomImpl.java
@@ -19,8 +19,8 @@ import com.codesquad.issueTracker.comment.domain.Comment;
 import com.codesquad.issueTracker.issue.application.dto.IssueCoverResponse;
 import com.codesquad.issueTracker.issue.domain.Issue;
 import com.codesquad.issueTracker.issue.domain.MainFilter;
-import com.codesquad.issueTracker.issue.presentation.dto.FilterCondition;
-import com.codesquad.issueTracker.issue.presentation.dto.SubFilterDetail;
+import com.codesquad.issueTracker.issue.application.dto.FilterCondition;
+import com.codesquad.issueTracker.issue.application.dto.SubFilterDetail;
 import com.codesquad.issueTracker.label.domain.AttachedLabel;
 import com.querydsl.core.types.Predicate;
 import com.querydsl.jpa.impl.JPAQuery;
